Deleted old comms systems and implemented io module
Added IoServerManager trait and implemented TcpServerManager.
